+++ eva/src/components/notifications.vue
@@ -1,6 +1,7 @@
 <template>
   <div
     style="position:absolute; right: 0;"
+    class="ma-3"
   >
     <template
       v-for="(item, index) in alerts"
@@ -10,7 +11,10 @@
         :key="`${index}/${item}`"
         :value="!!item"
         dismissible
+        dense
         type="warning"
+        class="mb-3"
+        :icon="item.icon || mdiSyncAlert"
         @input="dismissAllert(index)"
       >
         {{ item.message }}
@@ -19,8 +23,15 @@
   </div>
 </template>
 <script>
+import { mdiSyncAlert } from '@mdi/js';
+
 export default {
   name: 'Notifications',
+  data() {
+    return {
+      mdiSyncAlert,
+    };
+  },
   computed: {
     alerts: {
       get() {